refactor(p2): extract ToolCallPreviewSheet into sibling (stage 1 of 4)
Moves the tool-call preview ModalBottomSheet plus its content subviews
(ToolCallPreviewSheet, SearchWebPreview, ScrapeWebPreview, GenericToolPreview)
from ChatMessageTools.kt L903-L1238 into a new sibling
ChatMessageToolPreviewSheet.kt. Bodies are byte-equal with the parent commit
ae597a0d.

Widening:
  - sibling: ToolCallPreviewSheet private → internal (called from
    ChatMessageToolStep L893 in coordinator).
  - coord: getStringContent JsonElement?.getStringContent private → internal
    (referenced 3 times in extracted ToolPreview block; also needed by future
    SubAgent stage).
  - coord: ToolNames private → internal (referenced in extracted block;
    decl L158).
  - coord: MemoryActions private → internal (referenced in extracted block;
    decl L187).

Coordinator strip: 2605 → 2269 (-336 body) → 2254 (-15 import cleanup) = 2254.
Imports trimmed: 15 this-refactor orphans (Card, CardDefaults, IconButton,
TextStyle, TextAlign, sp, fastForEach, HighlightText, Delete01, MemoryRepository,
HighlightCodeBlock, Favicon, FormItem, JsonInstantPretty, openUrl).
Pre-existing dead imports preserved (CLAUDE.md §3): Box, BoxWithConstraints,
CircleShape, FilledTonalButton, FilterChip, LocalContentColor, Color, getValue,
setValue (the last two used implicitly via `by` delegation — 16 sites in
stripped coord).

Pre-cut discovery note: the pre-cut decl table missed top-level `private
object`/`enum class` decls (ToolNames, MemoryActions, AgentToolStatus,
AgentToolKind, ClipboardActions). Stages 3-4 will need AgentToolStatus +
AgentToolKind p→i as well.

Baseline 467 tests / 3 failed preserved.
